modify generate documentation script
